Get attachment
Getting attachment works, and we make use of the browser's built-in
caching capabilities to hit S3 with a GetObjectCommand as little as
possible. Something about this approach still doesn't go well with me.
I'll try Tigris' CloudFront-compatible cookie-based access next.

To do so we'll need to store things with paths
(/workspace/[id]/task/[id]/[attachment]), giving us the possibility to
generate cookies with folder-based access.