Drop workspace slug
Drop workspace slug until we convert workspaces into boards and
introduce orgs as workspaces. The idea behind unique slugs was that
they'd be unique to their organizations, which will also be identifiable
by a slug. Currently, this field + constraint just means that users
can't create workspaces with the same name.