Clean storage + Local MinIO
We are now using MinIO for local development. This makes the set up
easier to manage & inspect, reduces the amount of .env variables we
need, and reduces the risk of accidently deleting something from
production.

We are also moving some S3-specific business logic into a service, which
should make things cleaner and easier to migrate (if we ever need to) in
the future.